requirements.txt
django
djangorestframework
django-cors-headers
markdown
djangorestframework-simplejwt
openai
llama_index
langchain